% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale.R
\name{cc_location}
\alias{cc_location}
\alias{cc_macquarie}
\alias{cc_davis}
\alias{cc_mawson}
\alias{cc_casey}
\alias{cc_heard}
\alias{cc_kingston}
\title{Miscellaneous places}
\usage{
cc_location(loc = NULL, buffer = 5000, type = "mapbox.satellite",
  ..., debug = FALSE)

cc_macquarie(loc = c(158.93835, -54.49871), buffer = 5000,
  type = "mapbox.outdoors", ..., debug = FALSE)

cc_davis(loc = c(77 + 58/60 + 3/3600, -(68 + 34/60 + 36/3600)),
  buffer = 5000, type = "mapbox.outdoors", ..., debug = FALSE)

cc_mawson(loc = c(62 + 52/60 + 27/3600, -(67 + 36/60 + 12/3600)),
  buffer = 5000, type = "mapbox.outdoors", ..., debug = FALSE)

cc_casey(loc = cbind(110 + 31/60 + 36/3600, -(66 + 16/60 + 57/3600)),
  buffer = 5000, type = "mapbox.outdoors", ..., debug = FALSE)

cc_heard(loc = c(73 + 30/60 + 30/3600, -(53 + 0 + 0/3600)),
  buffer = 5000, type = "mapbox.outdoors", ..., debug = FALSE)

cc_kingston(loc = c(-147.70837, -42.98682), buffer = 5000,
  type = "mapbox.outdoors", ..., debug = FALSE)
}
\arguments{
\item{loc}{a longitude, latitude pair of coordinates}

\item{buffer}{with in metres to extend around the location}

\item{type}{character string of Mapbox service (see Details)}

\item{...}{dots, ignored currently}

\item{debug}{optionally print out files that will be used}
}
\value{
RasterBrick, with RGB
}
\description{
Visit some nice locales with web tiles.
}
\details{
Available types are 'mapbox.satellite', 'mapbox.outdoors', 'mapbox.terrain-rgb' but any string
accepted by Mapbox services will be passed through.
}
\examples{
## requres Mapbox key set in env var 'MAPBOX_API_KEY'
\dontrun{
im <- cc_macquarie()
library(raster)
plotRGB(im)
}
}
