% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceramic-package.R
\docType{package}
\name{ceramic-package}
\alias{ceramic}
\alias{ceramic-package}
\title{ceramic: Get imagery tiles}
\description{
Download imagery tiles to a standard cache and convert into raster objects. 
 Providers for 'AWS' terrain and 'Mapbox' are available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/ceramic}
  \item Report bugs at \url{https://github.com/hypertidy/ceramic/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (0000-0002-2471-7511)

Other contributors:
\itemize{
  \item Miles McBain (0000-0003-2865-2548) [contributor]
  \item Ben Raymond (regex wizardry) [contributor]
}

}
\keyword{internal}
