% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{ceramic_tiles}
\alias{ceramic_tiles}
\title{Tile files}
\usage{
ceramic_tiles(zoom = NULL, type = "mapbox.satellite",
  source = "api.mapbox.com", glob = "*.jpg*", regexp = NULL)
}
\arguments{
\item{zoom}{zoom level}

\item{type}{imagery type}

\item{source}{imagery source}

\item{glob}{see `fs::dir_ls`}

\item{regexp}{see `fs::dir_ls`}
}
\description{
Find existing files in the cache. Various options can be controlled, this is WIP
and due to change pending generalization across providers.
}
