% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceramic-package.R
\docType{data}
\name{cities}
\alias{cities}
\title{Cities locations}
\description{
Dataset from package {maps}.
}
\details{
Data frame with columns "name"        "country.etc" "pop"         "lat"         "long"        "capital".
}
