% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_ceramic_cache}
\alias{clear_ceramic_cache}
\title{Clear ceramic cache}
\usage{
clear_ceramic_cache(clobber = FALSE, ...)
}
\arguments{
\item{clobber}{set to \code{TRUE} to avoid checks and delete files}

\item{...}{reserved for future arguments, currently ignored}
}
\value{
This function is called for its side effect, but also returns the file paths as a
character vector whether deleted or not, or NULL if the user cancels.
}
\description{
Delete all downloaded files in the \code{\link[=ceramic_cache]{ceramic_cache()}}.
}
