% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference.R
\name{mercator_tile_extent}
\alias{mercator_tile_extent}
\title{Tile extent}
\usage{
mercator_tile_extent(tile_x, tile_y, zoom, tile_size = 256)
}
\arguments{
\item{tile_x}{x coordinate of tile}

\item{tile_y}{y coordinate of tile}

\item{zoom}{zoo level}

\item{tile_size}{tile dimensions (assumed square atm, 256*256)}
}
\description{
Calculate tile extent for a given x, y tile at a zoom level.
}
\details{
Currently only Mapbox spherical Mercator is supported.
}
