% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceramic-package.R
\docType{data}
\name{oz_laea}
\alias{oz_laea}
\title{Australian coastline}
\description{
The Australian coastline in simple features form, on a Lambert Azimuthal Equal Area projection.
}
\details{
This data set was derived from the 'oz' package.
}
