% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference.R
\name{plot_tiles}
\alias{plot_tiles}
\alias{tiles_to_polygon}
\title{Plot slipppy map tiles}
\usage{
plot_tiles(x, ..., add = FALSE, label = TRUE, cex = 0.6,
  add_coast = TRUE, include_zoom = TRUE)

tiles_to_polygon(x)
}
\arguments{
\item{x}{tiles as create by `ceramic_tiles()`}

\item{...}{arguments passed to `graphics::rect()`}

\item{add}{add to an existing plot?}

\item{label}{include text label?}

\item{cex}{relative size of text label if drawn (see `text()`)}

\item{add_coast}{include a basic coastline on the plot?}

\item{include_zoom}{include zoom level with text label if drawn?}
}
\description{
Create a new plot of tile rectangles, or add to an existing plot. The extent
('xmin', 'xmax', 'ymin', 'ymax') is used directly to draw the tiles so must be in the
native Mercator coordinate system used by most tile servers.
}
