% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{slippy_cache}
\alias{slippy_cache}
\title{Ceramic file cache}
\usage{
slippy_cache()
}
\value{
character value of location of cache
}
\description{
File system location where ceramic stores its cache.
}
\examples{
slippy_cache()
}
