% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiles.R
\name{unpack_rgb}
\alias{unpack_rgb}
\title{Unpack Mapbox terrain-RGB}
\usage{
unpack_rgb(x, filename = "")
}
\arguments{
\item{x}{three layer raster object}

\item{filename}{optional, filename to store the output}
}
\value{
terra rast object with one numeric layer
}
\description{
Mapbox terrain-rgb stores global elevation packed into Byte integers.
}
\details{
This function unpacks the three layers of a raster to give floating point elevation data.
}
\examples{
if (interactive() && !is.null(get_api_key())) {
unpack_rgb(read_tiles(type = "mapbox.terrain-rgb"))
}
}
