% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{ceramic_tiles}
\alias{ceramic_tiles}
\title{Tile files}
\usage{
ceramic_tiles(
  zoom = NULL,
  type = "mapbox.satellite",
  source = "api.mapbox.com",
  glob = NULL,
  regexp = NULL
)
}
\arguments{
\item{zoom}{zoom level}

\item{type}{imagery type}

\item{source}{imagery source}

\item{glob}{see \code{fs::dir_ls}}

\item{regexp}{see \code{fs::dir_ls}}
}
\value{
A data frame of tile file paths with tile index, zoom, type, version,
source and spatial extent.
}
\description{
Find existing files in the cache. Various options can be controlled, this is
liable to change pending generalization across providers.
}
\examples{
if (interactive() && !is.null(get_api_key())) {
 tiles <- ceramic_tiles(zoom = 0)
}
}
