% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{down_loader}
\alias{down_loader}
\title{Download tool for image tiles}
\usage{
down_loader(x, query_string, clobber = FALSE, ..., debug = FALSE)
}
\arguments{
\item{x}{tiles object}

\item{query_string}{an api query template (see Details)}

\item{clobber}{if `TRUE` re download any existing tiles}

\item{...}{ignored}

\item{debug}{simple debugging info printed if `TRUE`}
}
\value{
WIP
}
\description{
Tiles are cached with the native name of the source.
}
\details{
`query_string` takes the form of a template, see examples
}
